core_id values larger than CPU_SETSIZE
the /sys/devices/system/cpu/cpu*/topology/core_id values are not numbered
consecutively on Power, therefore the numbers might be higher than the
CPU_SETSIZE value. for now, just hack it; real fix is to make the code fully
understand and handle the oddities of the Power numa and cpu numbering.