AER-5897 - added 'size' parameter to 'cluster-stable' info command, and made 'namespace' parameter optional.
Additionally, it now checks that the cluster key didn't change while fetching
checked data.

Rationale: explaining how to accurately determine if migrations are complete is
too complicated. These changes allow the user to supply this command with a
target *size* and the server will verify that the target *size* matches the
current cluster size. Additionally, by not supplying the namespace the user is
now requesting that all namespaces have completed migrations.