cf_atomic_*_get API change, now take pointer
Atomic operations are atomic on a memory location. The APIs
in aerospike-common previously treated them as values.

To allow the reference model of C11 atomic operations into
the framework the cf_atomic_*_get functions have been change
to take a pointer.

A number of cases of ck_pr_load_* from the concurrency kit
have been replaces with the corresponding citrusleaf/common
implementation.